clean up source search
